backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
azure-ai-contentsafety==1.0.0
azure-core==1.39.0
azure-identity==1.25.3
azure-keyvault-secrets==4.10.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
ecdsa==0.19.1
fastapi==0.115.14
frozenlist==1.8.0
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
greenlet==3.3.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
isodate==0.7.2
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.0
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.20
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
marshmallow==3.26.2
msal==1.35.1
msal-extensions==1.3.1
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
openai==1.82.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.15
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
passlib==1.7.4
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pyodbc==5.2.0
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlglot==26.33.0
starlette==0.46.2
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.34.3
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
